Instate msgspec type registration mechanism
